Began refactoring deploy command
As part of another body of work, I needed to add support for another deployment "type". Upon inspecting the deployCharmOrBundle method, I found it was very difficult to follow the execution paths due to the high cyclomatic complexity. This refactoring is meant to address this.

It's not perfect or complete, but it's the 1st stopping point that works as intended.