Change how we cache 'last-revno' in memory.
Instead of just caching a watchKey => revno, break up the one-giant-map
into a bunch of maps, grouping them first by collection, and then when
possible by model-uuid. This should lead to lower memory consumption
because all of the common 'collection' strings are now deduped, and
we similarly remove all of the common model-uuid strings from most ids.
It has the cost of parsing them into and out of the map, and more maps
overall because of the separation. So testing and tweaking may be necessary.