Wire metrics up to the raftlease.FSM backend
This was quite a bit more invasive than I would have hoped for, but I
think it is correct.

One change is that we get rid of the distinction for
raftlease.OperationClientMetrics vs raftlease.clientMetrics. The only
difference was a simplification wrapper that made it an interface for
the operation but then thunked the values to the concrete type. Just do
that directly.

Implement FSMMetrics that does similar things but focuses on the backend
operations instead of the client side API requests.

The big thing that we wanted to add was just to include tracking for
when we call setTime and how many leases get expired. This requires
registering the metrics collector and cleaning it up, which has very
strange implications for the dependency engine (why isn't raftlease.FSM
a Worker?)
In general, metrics collectors registration and cleanup feels very ad
hoc. It feels like we should do something cleaner with DependencyEngine.
Maybe something where any given worker could conform to Collect and
Describe and the engine would register and unregister it directly?